TICKET_DASHBOARD/requirements.txt
aiosmtplib==3.0.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
fastapi-mail==1.5.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
pydantic==2.11.10
pydantic-core==2.33.2
pydantic-settings==2.11.0
pygments==2.19.2
pyjwt==2.9.0
pymongo==4.15.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.3
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
sentry-sdk==2.39.0
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
